📝 docs(tooltips): Simple icon with tooltip special case.
Since the problem is not related to the tooltips or the icons but the way how developers use them together, the proposed solution is not to create a new component but instead provide developers with an example on how to properly achieve the intended design. For that, the documentation of tooltips was updated providing on the “Special Cases” tab and “Case #3” all the required information for developers.